backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
qdrant-client==1.9.0
sqlalchemy==2.0.31
asyncpg==0.29.0
aiosqlite==0.20.0
alembic==1.13.2
psycopg2-binary==2.9.9
minio==7.2.7
python-docx==1.1.2
PyMuPDF==1.24.9
pdfplumber==0.11.4
pytesseract==0.3.10
Pillow==10.4.0
langchain-text-splitters==0.2.2
httpx==0.27.0
tenacity==8.5.0
pydantic==2.8.2
pydantic-settings==2.3.4
orjson==3.10.6
python-dotenv==1.0.1
loguru==0.7.2
numpy==1.26.4
scikit-learn==1.5.1
rapidfuzz==3.9.3
pytest==8.3.2
pytest-asyncio==0.23.8
requests==2.32.3
